Disable JetBrains AI Assistant by default
I use AOC mostly as a teaching resource. I’ve done this as it is unhelpful to have solutions suggested while discussing possible solutions. There is nothing wrong with using Generative AI to assist writing code, when it is done in a deliberate informed way. I don’t believe it is particularly helpful when the objective is learning.